requirements.txt
atproto>=0.0.61
anthropic>=0.58.2
click>=8.2.1
python-dotenv>=1.1.1
pydantic>=2.11.7
rich>=14.0.0

# Development dependencies
pytest>=7.0.0
pytest-mock>=3.0.0
